requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
numpy>=1.24.0
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.15.0
tqdm>=4.65.0
tensorboard>=2.12.0
librosa>=0.10.0